ADT Launch: default to always showing device chooser dialog. do not merge.
The device chooser dialog has a checkbox that says use the same
device for future launches of an application. This is currently
enabled, which does not match current user expectation that the
dialog is always shown.

This patch disables the checkbox by default, so that the behavior
is identical to previous versions of ADT.

(cherry picked from commit b8baa65a172fc38aef0b24a6c66f5ed42bf10b39)

Change-Id: I6831c4088281f67ad9b0c5f75a30b8cfbc9f2bdf