Fix auto-indent for text nodes after closing tag (issue #22332). do not merge.
This changeset fixes the auto-indent handling for a specific scenario:
pressing newline within a text node, when the next non-text region is
an element end tag, and the caret is not on the same line as the end
tag. This scenario is covered by the new unit tests testIssue22332a
and testIssue22332b.

(cherry picked from commit 45fa942e9b5c773c72b703194bc0470e991a4a0f)

Change-Id: Id2b39e428a48c8d9df2e4c9b7e304968fe42d19b