Make Fix in Lint Warnings window work for closed files
When you click on the Fix icon in the Lint Warnings window, the marker
you're applying a fix for may be referencing a file that is not open
and visible. The apply fix code needs to account for this, and open
the document (such that getModelForEdit will succeed) and for
interactive fixes, also open the editor.

(cherry picked from commit 9b5ee8cc697ab75826f7c53718b84042fa45cf74)

Change-Id: I42cf1f853d05686cd34c3c5c6918c44b95da5056