logcat: fix npe & clear log in filtered view. do not merge.
1. Fix NPE when selecting all items: Issue is that in case of a
table with SWT.VIRTUAL, tableItem.getData() could return NULL if
the item hasn't been shown. So we have to rely on the original
model to retrieve the items.

2. When viewing items after applying a filter, clear log will
clear the full log but not the filtered view. To clear the filtered
view, the SWT ViewerFilters have to be reapplied.

(cherry picked from commit 4481c9d73bcbb7afb958a575d1d86dd2e2c21719)

Change-Id: I8df89b4bc3bcd03e66e5769f4f8de1fba8c409ac