logcat: when copying items, index into filtered list. do not merge.
The previous commit introduced a bug: The SWT table model is a list of
logcat messages before any filters are applied. But the user sees the
filtered messages. So when user selects items in the table, the
selection index will not match the index in the input model.

This patch applies all current filters to the input model and then
indexes this filtered list to get the same items that the user has
selected in the view.

(cherry picked from commit daa5a5c9200708b09983bf3852b94bb8ebab1e27)

Change-Id: I0120555f621c47e9ae46522f9c4c13ac1ad36cff