Fix Run Lint toolbar action to pick up more selection contexts. do not merge.
You currently have to select a project in the package explorer before
you click on the Run Lint toolbar action. This changeset makes the
action look harder for which project to run lint for:

- As before, look for the current selection and check whether it
  provides a selection
- If there's an active editor, use its project.
- If not, if there's just one available Android project, use it.
- If not, if there's just one non-library Android project, use it.

Finally, display a warning dialog if no project can be found, or if
the selected project is not an Android project, such that it's
apparent why nothing happens.

(cherry picked from commit bc32af8dc9e20135c29ba7444ae162761774cdb5)

Change-Id: Iaba5a6bfeeabe1f27c127bd08db832470da8f816