Provide readable device name rather than just serial number
There are a number of places where we need to display a device
name. Currently, DDMS shows the serial number + manufacturer/model
if available, but in other instances, we just show the serial
number.

This CL provides a getName() method in the IDevice interface so
that all the locations which require a human readable name can
use this.

The name is typically of the form manufacturer-model-serial.
If the manufacturer/model strings are not available, then those
fields are omitted, so it will show just the serial number.

(cherry picked from commit b7f70835c533f0f07a1f52f75a1bbe783c32e010)

Change-Id: Ifffbc146050d7fb54a4da811a55e1f3e02d4a2e6