Fix emma for split projects.
The location of the .em file was wrong. Changing its location
means it gets put in the tested project so we need a way to know where
the output of the project is.

As we need to be better at supporting customized projects folder, we
introduce a new task that will be able to look for the customized
path of any projects. So far we only support the output folder
and the source folder.

(cherry picked from commit fcb82e3185d657e3b5440295478cdc58afd168c3)

Change-Id: Ifab65082aae8753ff0fc01be9f6337d5032edfa0