Fix issue where libraries were not getting instrumented.
They actually were instrumented but the emma output file was not
passed to the report task.

After trying to make each library generate its own .em file and
passing to the report task, I gave up due to FileSet (and the
report task) being stupid.

So the main fix is having the libraries contribute to the main coverage.em
file (since it's build using the merge option) which location is passed
to the library project.

The changes to the custom task was to support querying the libraries for
a path with a given leaf (from the lib root folder) but this is not
used at the moment. The other tasks are slightly cleaned up due to
this though and are getting prepared to support customized folder
names in folder.

(cherry picked from commit aaeb6a3d58ce5dccc91360bb429b5e37af650619)

Change-Id: I3d6fe39825f00e3d43e5f136f4943a2f27d0aa92