Preview support in the templates
This changeset adds a Preview page to the end of the template wizards
which shows the changes to be applied to the project - first the files
which were edited (merged), such as the manifest file in the case of a
new activity, and second the text files to be created, and third the
binary files to be created.

In addition, the user can now uncheck any of these changes, if for
example the manifest file edit isn't wanted. Furthermore, the now that
the changes are computed up front, the phase of applying the changes
is run in the background with a progress bar in the New-wizard.
There's also some consolidation of the various template wizards to
handle things like the upgrade-page and the dependency-page in one
place.

Infrastructure wise this changes the template instantiation from being
based on File manipulation to using the refactoring infrastructure's
change support, which should be more reliable. It also fixes a bug
where projects were marked as library projects which should not have
be.

(cherry picked from commit 802de810020fba3f86282cd1d66597a2a41698e3)

Change-Id: Ia1321e86a226d36c8b60032b5566a79236486e94