Fix validation where selecting same build and min sdk version
Fix bug where selecting the same min sdk version and build target
version would result in a wizard complaint that the minSdkVersion
was higher than the build target API level. This happened because
the code passed the minSdk string to the AndroidVersion compare
method, which assumes that string is a codename; in the wizard
case it can be the string representation of the api level.

(cherry picked from commit 86b9ea11f0f65ac33c417712b2f50022c5b84977)

Change-Id: Ie6affec980dd1632e972e9e3a54dbb5f5ea45c73